+++ jest.config.js
@@ -1,0 +1,5 @@
+module.exports = {
+  clearMocks: true,
+  testMatch: ['**/*.spec.ts'],
+  preset: 'ts-jest',
+};